Restructure project tree: add components folder and move created components into folder
